requirements_new.txt
docling==2.0.0
streamlit==1.30.0
paddleocr==2.7.3
pytesseract==0.3.13
pillow==10.0.1
pandas==2.0.3
numpy==1.24.3
opencv-python==4.8.1.78
pdf2image==1.16.3
torch==2.1.0
transformers==4.34.0